refactored login and signup
