+++ app/utilities/AdminUtilities.js
@@ -1,0 +1,10 @@
+import db from '../lib/firebase';
+import { doc, collection, getDoc, getDocs, updateDoc, setDoc } from 'firebase/firestore';
+
+
+
+export async function getAdminDoc(uid){
+    const adminDoc = doc(db, 'admin', uid);
+    const adminSnapshot = await getDoc(admin);
+    return adminSnapshot;
+}
